backend/requirements.txt
fastapi
uvicorn
sqlalchemy
requests
beautifulsoup4
readability-lxml
